fix lazy list key error on android

+++ composeApp/src/commonMain/kotlin/com/github/singularity/ui/feature/settings/SettingsScreen.kt
@@ -133,7 +133,7 @@ private fun SettingsScreen(
                 ) {
                     items(
                         items = ColorsList,
-                        key = { it.value },
+	                    key = { it.value.toLong() },
                         contentType = { Color::class }
                     ) {
                         Box(
